backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=1.6.0
langchain==0.0.350
pinecone-client==2.2.4
python-jose==3.3.0
stripe==7.0.0
Pillow==10.1.0
requests==2.31.0
zeep==4.2.1  # Added for SanMar SOAP API